- moved Note resource into own module - added directories
