+++ core/util/constants.py
@@ -1,0 +1,9 @@
+NOTE_BODY_CHAR_LIMIT = 250
+NOTE_TITLE_CHAR_LIMIT = 30
+
+# HTTP response codes
+OK = 200
+CREATED = 201
+NO_CONTENT = 204
+BAD_REQUEST = 400
+NOT_FOUND = 404
